+++ include/mbgl/route/route.hpp
@@ -1,0 +1,18 @@
+
+#pragma once
+
+#include <mbgl/util/geometry.hpp>
+#include <mbgl/util/color.hpp>
+#include <string>
+
+namespace mbgl {
+namespace route {
+
+class Route {
+    public:
+    Route(const std::string& route);
+
+};
+
+} // namespace gfx
+} // namespace mbgl
