+++ src/mbgl/route/route.cpp
@@ -1,0 +1,9 @@
+
+#include <mbgl/route/route_options.hpp>
+
+namespace mbgl {
+
+namespace route {
+
+}
+}
