--- src/mbgl/route/route_options.cpp
@@ -1,10 +1,0 @@
-
-#include <mbgl/route/route_options.hpp>
-
-namespace mbgl {
-
-namespace route {
-const std::string RouteSegmentOptions::BASE_ROUTE_SEGMENT_STR = "base_route_segment";
-
-}
-}
